did: merge things exclusive to date_picker namespace
